+++ packages/forms/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @under-control/forms
 
+## 1.6.16
+
+### Patch Changes
+
+- Add merged bind function
+- Updated dependencies
+  - @under-control/inputs@1.6.16
+  - @under-control/core@1.6.16
+  - @under-control/validate@1.6.16
+
 ## 1.6.15
 
 ### Patch Changes