+++ packages/validate/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @under-control/validate
 
+## 1.6.16
+
+### Patch Changes
+
+- Add merged bind function
+- Updated dependencies
+  - @under-control/core@1.6.16
+
 ## 1.6.15
 
 ### Patch Changes